User supplied patches + enhancements for 2nd encoding in auto mode
git-svn-id: https://svn.code.sf.net/p/kmttg/code/trunk@1891 b0df3fdd-2057-4f73-b226-dfc9190e8f7a

+++ src/com/tivo/kmttg/task/encode.java
@@ -230,7 +230,9 @@ public class encode implements Serializable {
             }
             
             // Remove .mpg file if option enabled
-            if (config.RemoveMpegFile == 1) {
+            // If there is a second encode job working off the same source file,
+            // then the source should not be removed.
+            if (!job.hasMoreEncodingJobs && config.RemoveMpegFile == 1) {
                if ( file.delete(job.inputFile) ) {
                   log.print("(Deleted file: " + job.inputFile + ")");
                } else {
